requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
black==26.3.1
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
comm==0.2.3
configmypy==0.2.0
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
einops==0.8.2
executing==2.2.1
filelock==3.25.2
fonttools==4.62.1
fsspec==2024.10.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
h5py==3.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
idna==3.11
ImageIO==2.37.3
iniconfig==2.3.0
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
neuraloperator==0.3.0
numpy==2.4.3
opt_einsum==3.4.0
packaging==26.0
parso==0.8.6
pathspec==1.0.4
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
pytokens==0.4.1
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
rich==14.3.3
ruamel.yaml==0.19.1
safetensors==0.7.0
scipy==1.17.1
sentry-sdk==2.54.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
stack-data==0.6.3
sympy==1.14.0
tensorly==0.9.0
tensorly-torch==0.5.0
the_well==1.2.0
timm==1.0.25
torch==2.10.0
torch_harmonics==0.8.0
torchvision==0.25.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wandb==0.25.1
wcwidth==0.6.0
zencfg==0.7.0
